+++ TelegramBotAPI-extensions-api/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/extensions/api/send/media/SendDocument.kt
@@ -1,8 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.extensions.api.send.media
 
 import com.github.insanusmokrassar.TelegramBotAPI.bot.RequestsExecutor
-import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.FileId
-import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.MultipartFile
+import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.*
 import com.github.insanusmokrassar.TelegramBotAPI.requests.send.media.SendDocument
 import com.github.insanusmokrassar.TelegramBotAPI.types.ChatIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
@@ -13,8 +12,8 @@ import com.github.insanusmokrassar.TelegramBotAPI.types.files.PhotoSize
 
 suspend fun RequestsExecutor.sendDocument(
     chatId: ChatIdentifier,
-    document: FileId,
-    thumb: FileId? = null,
+    document: InputFile,
+    thumb: InputFile? = null,
     text: String? = null,
     parseMode: ParseMode? = null,
     disableNotification: Boolean = false,
@@ -36,104 +35,11 @@ suspend fun RequestsExecutor.sendDocument(
 suspend fun RequestsExecutor.sendDocument(
     chatId: ChatIdentifier,
     document: DocumentFile,
-    thumb: PhotoSize? = document.thumb,
     text: String? = null,
     parseMode: ParseMode? = null,
     disableNotification: Boolean = false,
     replyToMessageId: MessageIdentifier? = null,
     replyMarkup: KeyboardMarkup? = null
 ) = sendDocument(
-    chatId, document.fileId, thumb ?.fileId, text, parseMode, disableNotification, replyToMessageId, replyMarkup
-)
-
-suspend fun RequestsExecutor.sendDocument(
-    chatId: ChatIdentifier,
-    document: MultipartFile,
-    thumb: FileId? = null,
-    text: String? = null,
-    parseMode: ParseMode? = null,
-    disableNotification: Boolean = false,
-    replyToMessageId: MessageIdentifier? = null,
-    replyMarkup: KeyboardMarkup? = null
-) = execute(
-    SendDocument(
-        chatId,
-        document,
-        thumb,
-        text,
-        parseMode,
-        disableNotification,
-        replyToMessageId,
-        replyMarkup
-    )
-)
-
-suspend fun RequestsExecutor.sendDocument(
-    chatId: ChatIdentifier,
-    document: MultipartFile,
-    thumb: MultipartFile? = null,
-    text: String? = null,
-    parseMode: ParseMode? = null,
-    disableNotification: Boolean = false,
-    replyToMessageId: MessageIdentifier? = null,
-    replyMarkup: KeyboardMarkup? = null
-) = execute(
-    SendDocument(
-        chatId,
-        document,
-        thumb,
-        text,
-        parseMode,
-        disableNotification,
-        replyToMessageId,
-        replyMarkup
-    )
-)
-
-suspend fun RequestsExecutor.sendDocument(
-    chatId: ChatIdentifier,
-    document: FileId,
-    thumb: MultipartFile,
-    text: String? = null,
-    parseMode: ParseMode? = null,
-    disableNotification: Boolean = false,
-    replyToMessageId: MessageIdentifier? = null,
-    replyMarkup: KeyboardMarkup? = null
-) = execute(
-    SendDocument(
-        chatId,
-        document,
-        thumb,
-        text,
-        parseMode,
-        disableNotification,
-        replyToMessageId,
-        replyMarkup
-    )
-)
-
-suspend fun RequestsExecutor.sendDocument(
-    chatId: ChatIdentifier,
-    document: MultipartFile,
-    thumb: PhotoSize? = null,
-    text: String? = null,
-    parseMode: ParseMode? = null,
-    disableNotification: Boolean = false,
-    replyToMessageId: MessageIdentifier? = null,
-    replyMarkup: KeyboardMarkup? = null
-) = sendDocument(
-    chatId, document, thumb ?.fileId , text, parseMode, disableNotification, replyToMessageId, replyMarkup
-)
-
-suspend fun RequestsExecutor.sendDocument(
-    chatId: ChatIdentifier,
-    document: DocumentFile,
-    thumb: MultipartFile,
-    text: String? = null,
-    parseMode: ParseMode? = null,
-    disableNotification: Boolean = false,
-    replyToMessageId: MessageIdentifier? = null,
-    replyMarkup: KeyboardMarkup? = null
-) = sendDocument(
-    chatId, document.fileId, thumb, text, parseMode, disableNotification, replyToMessageId, replyMarkup
+    chatId, document.fileId, document.thumb ?.fileId, text, parseMode, disableNotification, replyToMessageId, replyMarkup
 )