+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/DeleteMessage.kt
@@ -23,26 +23,3 @@ data class DeleteMessage(
     override val requestSerializer: SerializationStrategy<*>
         get() = serializer()
 }
-
-@Deprecated("Deprecated due to extracting into separated library")
-suspend fun RequestsExecutor.deleteMessage(
-    chatId: ChatIdentifier,
-    messageId: MessageIdentifier
-) = execute(
-    DeleteMessage(chatId, messageId)
-)
-
-@Deprecated("Deprecated due to extracting into separated library")
-suspend fun RequestsExecutor.deleteMessage(
-    chat: Chat,
-    messageId: MessageIdentifier
-) = deleteMessage(chat.id, messageId)
-
-@Deprecated("Deprecated due to extracting into separated library")
-suspend fun RequestsExecutor.deleteMessage(
-    message: Message
-) = deleteMessage(message.chat, message.messageId)
-
-suspend fun Message.delete(
-    requestsExecutor: RequestsExecutor
-) = requestsExecutor.deleteMessage(this)