+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/chat/ExportChatInviteLink.kt
@@ -20,13 +20,3 @@ data class ExportChatInviteLink(
     override val requestSerializer: SerializationStrategy<*>
         get() = serializer()
 }
-
-@Deprecated("Deprecated due to extracting into separated library")
-suspend fun RequestsExecutor.exportChatInviteLink(
-    chatId: ChatIdentifier
-) = execute(ExportChatInviteLink(chatId))
-
-@Deprecated("Deprecated due to extracting into separated library")
-suspend fun RequestsExecutor.exportChatInviteLink(
-    chat: PublicChat
-) = exportChatInviteLink(chat.id)