+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/chat/members/UnbanChatMember.kt
@@ -20,28 +20,3 @@ data class UnbanChatMember(
     override val requestSerializer: SerializationStrategy<*>
         get() = serializer()
 }
-
-@Deprecated("Deprecated due to extracting into separated library")
-suspend fun RequestsExecutor.unbanChatMember(
-    chatId: ChatIdentifier,
-    userId: UserId
-) = execute(UnbanChatMember(chatId, userId))
-
-@Deprecated("Deprecated due to extracting into separated library")
-suspend fun RequestsExecutor.unbanChatMember(
-    chat: PublicChat,
-    userId: UserId
-) = unbanChatMember(chat.id, userId)
-
-@Deprecated("Deprecated due to extracting into separated library")
-suspend fun RequestsExecutor.unbanChatMember(
-    chatId: ChatId,
-    user: User
-) = unbanChatMember(chatId, user.id)
-
-@Deprecated("Deprecated due to extracting into separated library")
-suspend fun RequestsExecutor.unbanChatMember(
-    chat: PublicChat,
-    user: User
-) = unbanChatMember(chat.id, user.id)
-