+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/stickers/DeleteStickerFromSet.kt
@@ -19,21 +19,3 @@ data class DeleteStickerFromSet(
     override val requestSerializer: SerializationStrategy<*>
         get() = serializer()
 }
-
-
-
-@Deprecated("Deprecated due to extracting into separated library")
-suspend fun RequestsExecutor.deleteStickerFromSet(
-    sticker: FileId
-) = execute(
-    DeleteStickerFromSet(
-        sticker
-    )
-)
-
-@Deprecated("Deprecated due to extracting into separated library")
-suspend fun RequestsExecutor.deleteStickerFromSet(
-    sticker: Sticker
-) = deleteStickerFromSet(
-    sticker.fileId
-)