+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # TelegramBotAPI changelog
 
+## 0.24.0
+
+* `TelegramBotAPI`:
+    * All suspend functions for `RequestsExecutor` was removed (due to replacement into
+    [TelegramBotAPI extensions project](TelegramBotAPI-extensions-api/README.md))
+    * `ForwardFromChannelInfo#channelChat` now is `ChannelChat` instead of `Chat`
+* `TelegramBotAPI-extensions-api`:
+    * Most part of sending media messages functions was removed and replaced with their `InputFile` args analogs
+
+
 ## 0.23.0 TelegramBotAPI 4.6
 
 * `Poll` now is sealed class